originally posted on gamedev stackoverflow 15th april 2019
I built a directx chunk from vertex bindings to limit the number of buffers used to update the instanced chunks vertices positions. I do not think it is a viable option for a standard voxel chunk basic example as each vertices are drawn even though the bytes/ints that form the chunk are supposed to be creating empty faces. I tried using a geometry shader to remove extra vertices when they are not supposed to be visible, but it wasn't working last time i tried. Still, it works, the program loads, but it lags due to too many vertices. I'm gonna have to create an alternative version at some point.

https://gamedev.stackexchange.com/questions/169506/vertex-manipulation-in-instanced-shader-hlsl-sharpdx-4-2?noredirect=1#comment302452_169506